feat(transfer): Use tranfer API instead of CLI
